% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_counts_list_to_tibble.R
\name{acs_counts_list_to_tibble}
\alias{acs_counts_list_to_tibble}
\title{Organize a list output from a JSON parser to a tibble of ACS counts data.}
\usage{
acs_counts_list_to_tibble(.list, table_name)
}
\arguments{
\item{.list}{A list that comes from parsing a rectangular ACS json array}

\item{table_name}{The name of the ACS table}
}
\value{
A tibble of pivoted, R-friendly data, with sub-totals removed.
}
\description{
Organize a list output from a JSON parser to a tibble of ACS counts data.
}
