% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_counts_pivot_longer.R
\name{acs_counts_pivot_longer}
\alias{acs_counts_pivot_longer}
\title{Split an ACS "Variable" value into three components}
\usage{
acs_counts_pivot_longer(.x, table_name)
}
\arguments{
\item{.x}{A data frame with many variables that contain \code{table_name}}

\item{table_name}{The name of the ACS table that the data come from}
}
\value{
A data frame where the "Variable" value has been split
}
\description{
Split an ACS "Variable" value into three components
}
