% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_counts_url.R
\name{acs_counts_url}
\alias{acs_counts_url}
\title{Build the URL for the JSON data file for an ACS summary table.}
\usage{
acs_counts_url(year, vintage, table_name, state_fips)
}
\arguments{
\item{year}{int, The reference year for the survey product}

\item{vintage}{int, Denotes the time-span the data are averaged over}

\item{table_name}{char, The full name of the ACS table, e.g. "C23002A"}

\item{state_fips}{int, The FIPS code for a state, e.g. 55 for Wisconsin.}
}
\value{
string, The URL of a JSON file on the Census's website
}
\description{
The fourth argument, "state_fips," must ultimately be replaced with a string
that captures the geographical part of the API query.
}
\examples{
acs_counts_url(2020, 5, "C23002A", 55)
}
