% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_groups_list_to_tibble.R
\name{acs_groups_list_to_tibble}
\alias{acs_groups_list_to_tibble}
\title{Organize a list output from a JSON parser to a tibble of ACS groups data.}
\usage{
acs_groups_list_to_tibble(.x, detail_names = NULL)
}
\arguments{
\item{.x}{A list that comes from parsing a rectangular ACS json array}

\item{detail_names}{The names of the table's grouping fields}
}
\value{
A tibble of pivoted, R-friendly metadata without annotation fields
}
\description{
Organize a list output from a JSON parser to a tibble of ACS groups data.
}
