% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_groups_url.R
\name{acs_groups_url}
\alias{acs_groups_url}
\title{Build the URL for the JSON metadata file for an ACS summary table}
\usage{
acs_groups_url(year, acs_type, table_name)
}
\arguments{
\item{year}{int, The reference year for the survey product}

\item{acs_type}{int, Denotes the time-span the data are averaged over}

\item{table_name}{char, An ACS summary table's full name, e.g. "C23002A".}
}
\value{
char : The URL of a JSON file on the Census's website.
}
\description{
Build the URL for the JSON metadata file for an ACS summary table
}
\examples{
acs_groups_url(2020, 5, "C23002A")
}
