% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_path.R
\name{acs_path}
\alias{acs_path}
\title{Check for valid options and return the path part of an API call for ACS data.}
\usage{
acs_path(.year, .year_span)
}
\arguments{
\item{.year}{An integer year between 2004 and the current year, inclusive.}

\item{.year_span}{Either 1, 3, or 5, depending upon the desired time resolution.}
}
\value{
a slash-separated URL path.
}
\description{
Check for valid options and return the path part of an API call for ACS data.
}
