% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ACS_RACE_ETHNICITY_VARIABLES}
\alias{ACS_RACE_ETHNICITY_VARIABLES}
\title{Variables for querying populations by race and ethnicity}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 15 rows and 4 columns.
}
\usage{
ACS_RACE_ETHNICITY_VARIABLES
}
\description{
Variables for querying populations by race and ethnicity
}
\keyword{datasets}
