% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key_utilities.R
\name{api_key_is_set}
\alias{api_key_is_set}
\title{Check if the local environment contains an expected constant defining an API key}
\usage{
api_key_is_set()
}
\value{
TRUE if the key exists, FALSE otherwise
}
\description{
Check if the local environment contains an expected constant defining an API key
}
\examples{
api_key_is_set()

}
\concept{API key}
