% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key_utilities.R
\name{api_key_setup}
\alias{api_key_setup}
\title{Guide a new user at setting up a Census API key}
\usage{
api_key_setup()
}
\value{
a numeric status: 0 for success, 1 for quitting, 2 if the key exists.
}
\description{
This only works in interactive mode. It checks if a key already exists in the
environment. The key corresponds to \code{Sys.getenv("CENSUS_API_KEY")} in order
to be compatible with the
\href{https://walker-data.com/tidycensus/reference/census_api_key.html}{tidycensus}
package. If the key doesn't exist then it prompts you to either go to the
Census website (it will open a browser window for you) or to enter in your
API key (by typing or pasting). It then sets the value in your current
environment and saves it to the .Renviron file in the user's HOME directory.
}
\examples{
api_key_setup()

}
\seealso{
\code{\link[tidycensus:census_api_key]{tidycensus::census_api_key()}}
}
\concept{API key}
