% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key_utilities.R
\name{api_key_value}
\alias{api_key_value}
\title{Get the current value of the local Census API key}
\usage{
api_key_value()
}
\value{
a hexadecimal API token, or an empty string
}
\description{
This corresponds to \code{Sys.getenv("CENSUS_API_KEY")} in order to be compatible
with the \href{https://walker-data.com/tidycensus/reference/census_api_key.html}{tidycensus}
package.
}
\examples{
api_key_value()

}
\seealso{
\code{\link[tidycensus:census_api_key]{tidycensus::census_api_key()}}
}
\concept{API key}
