% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_query.R
\name{api_query}
\alias{api_query}
\title{Construct the query part of the URL for a call to the US Census API}
\usage{
api_query(variables, for_geo, for_items, ..., use_key = TRUE)
}
\arguments{
\item{variables}{a vector of variable names, like \code{"B01001_001E"}}

\item{for_geo}{the geographical level the data will describe, e.g. \code{"tract"}}

\item{for_items}{the specific instances of \code{for_geo} desired, e.g. \code{"*"} or \code{"000200"}}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic dots}}> other items to pass to the query}

\item{use_key}{optional, should the query include a Census API key from the system environment. Defaults to \code{TRUE}}
}
\value{
a string with the query part of a Census API call
}
\description{
Construct the query part of the URL for a call to the US Census API
}
\examples{
api_query(paste0("B25003_00", 1:3, "E"),
                 "tract",
                 "*",
                 state = 55L,
                 county = 101L,
                 use_key = FALSE)

api_query(paste0("P1_00", c(1, 3, 4), "N"),
                 "block\%20group",
                 "*",
                 state = 55L,
                 county = 101L,
                 use_key = FALSE)
}
