% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{B21005_VARIABLES}
\alias{B21005_VARIABLES}
\title{Variables in B21005: Age x Veteran x Employed for Civilians 18-64}
\format{
A tibble with seven columns
\describe{
\item{Name}{The name of the variable, e.g. B21005_0001E}
\item{Age}{The Census's description of the age range described by this variable.}
\item{Veteran}{\code{TRUE} if the variable describes veterans, and \code{NA} for subtotals.}
\item{In Labor Force}{\code{TRUE}, \code{FALSE}, or \code{NA} for subtotals.}
\item{Employed}{\code{TRUE}, \code{FALSE}, or \code{NA} for subtotals.}
\item{Lower Age}{The lowest age that this variable describes.}
\item{Upper Age}{The highest age that this variable describes.}
}
}
\usage{
B21005_VARIABLES
}
\description{
Variables in B21005: Age x Veteran x Employed for Civilians 18-64
}
\keyword{datasets}
