% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_api_variable.R
\name{build_api_variable}
\alias{build_api_variable}
\title{Vectorized Creation of Census ACS Variables}
\usage{
build_api_variable(
  table_code,
  race_code,
  item_number,
  separator = "_",
  suffix = "E"
)
}
\arguments{
\item{table_code}{the table code, like "B18101"}

\item{race_code}{either an empty string or one of [A..I]}

\item{item_number}{some integer between 0 and 999}

\item{separator}{optional, usually "_" for ACS variables.}

\item{suffix}{optional, usually "E" for ACS variables, but sometimes "N" for decennial data.}
}
\value{
a vector of strings, each of which is a variable name
}
\description{
All detail table variables have the format "B000OOX_000E", where the "B" and
"E" are literal and the "X" is a race code.
}
\examples{
tables <- c("B19013", "B18101", "B18101")
races <- c("", "", "I")
numbers <- 1
build_api_variable(tables, races, numbers)
}
