% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_for_geographies.R
\name{build_for_geographies}
\alias{build_for_geographies}
\title{Use the \code{for} geography to specify the level that you are querying data for.}
\usage{
build_for_geographies(for_geo, ...)
}
\arguments{
\item{for_geo}{<chr> the geographical level the data will describe, e.g. \code{"tract"}}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> The specific items to search for, which will be all items if you leave them empty.}
}
\value{
a "key:value" string
}
\description{
Use the \code{for} geography to specify the level that you are querying data for.
}
\examples{
hercacstables:::build_for_geographies("block")
hercacstables:::build_for_geographies("tract", "000400", "000500")
}
\keyword{internal}
