% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_in_geographies.R
\name{build_in_geographies}
\alias{build_in_geographies}
\title{Create the part of a Census API query that describes containing geographies}
\usage{
build_in_geographies(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> key-value pairs like "state='03'"}
}
\value{
A vector of "geo:code" pairs
}
\description{
Create the part of a Census API query that describes containing geographies
}
\examples{
hercacstables:::build_in_geographies(state=55, county = 101, barf=NULL)

}
\keyword{internal}
