% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_query_parameters.R
\name{build_query_parameters}
\alias{build_query_parameters}
\title{Create the query part of a call to the Census's data API}
\usage{
build_query_parameters(
  variables,
  for_geo,
  for_items = NULL,
  in_geos = NULL,
  use_key = FALSE
)
}
\arguments{
\item{variables}{<chr[]> a vector of variable names, like \code{"B01001_001E"}}

\item{for_geo}{<chr> the geographical level the data will describe, e.g. \code{"tract"}}

\item{for_items}{<chr[]> one or more instances of \code{for_geo} desired, e.g. \code{"*"} or \code{"000200"}, passed on to \code{\link[=build_for_geographies]{build_for_geographies()}}}

\item{in_geos}{<lst> a list of key-value pairs to pass to \code{\link[=build_in_geographies]{build_in_geographies()}}}

\item{use_key}{<lgl?> optional, should the query include a Census API key from the system environment. Defaults to \code{TRUE}}
}
\value{
a list of key-value pairs that can be converted into a URL query.
}
\description{
Create the query part of a call to the Census's data API
}
\keyword{internal}
