% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c23002a_groups}
\alias{c23002a_groups}
\title{Groups available for breaking down ACS employment data by sex and age}
\format{
A list with 5 items:
\describe{
\item{url}{the actual URL used to query the Census's data API}
\item{json}{The JSON-formatted string that the Census responded with}
\item{list}{The list produced by \code{jsonlite::parse_json()}}
\item{fields}{The fields that \code{acs_groups_list_to_tibble} uses}
\item{data}{A data frame produced by \code{acs_groups_list_to_tibble()}}
}
}
\source{
https://www.census.gov/data/developers/guidance/api-user-guide.html
}
\usage{
c23002a_groups
}
\description{
A list containing several stages of parsing results from querying the US
Census Bureau's data API for the American Community Survey. This set contains
the results from searching for the summary table called "C23002A". That table
contains counts for white people broken down by sex, age, and employment
status.
}
\keyword{datasets}
