% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CENSUS_API_URL}
\alias{CENSUS_API_URL}
\title{The current URL for accessing the U.S. Census Bureau's API.}
\format{
An object of class \code{character} of length 1.
}
\source{
\link{https://api.census.gov/data}
}
\usage{
CENSUS_API_URL
}
\description{
The current URL for accessing the U.S. Census Bureau's API.
}
\keyword{datasets}
