% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CHILDREN_PER_FAMILY_METADATA}
\alias{CHILDREN_PER_FAMILY_METADATA}
\title{Categorize ACS variables about children per family}
\format{
\subsection{CHILDREN_PER_FAMILY_METADATA}{

A data frame with 15 rows and 6 columns
\describe{
\item{Group}{\if{html}{\out{<chr>}} The table, always "B09002"}
\item{Variable}{\if{html}{\out{<chr>}} The full variable name, e.g. "B09002_001E"}
\item{Index}{\if{html}{\out{<int>}} The row of this variable in the table}
\item{Adults}{\if{html}{\out{<chr>}} The adult(s) heading the household}
\item{Lower Age}{\if{html}{\out{<int>}} The age of the youngest children counted by this variable}
\item{Upper Age}{\if{html}{\out{<int>}} The age of the oldest children counted by this variable}
}
}
}
\source{
https://api.census.gov/data/2022/acs/acs1/groups/B09002.html
}
\usage{
CHILDREN_PER_FAMILY_METADATA
}
\description{
These data come from table B17026, "OWN CHILDREN UNDER 18 YEARS BY FAMILY TYPE AND AGE"
}
\keyword{datasets}
