% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DECENNIAL_POPULATION_FIELDS}
\alias{DECENNIAL_POPULATION_FIELDS}
\title{Variables from the Decennial censuses}
\format{
A tibble with five columns
\describe{
\item{Year}{The year of the decennial census.}
\item{Group}{The table that the variables come from.}
\item{Race/Ethnicity}{The OMB label for a race/ethnic group.}
\item{Index}{The row in the table that corresponds to the race/ethnicity.}
\item{Variable}{The name of the variable for an API call}
}
}
\usage{
DECENNIAL_POPULATION_FIELDS
}
\description{
Variables from the Decennial censuses
}
\keyword{datasets}
