% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EDUCATIONAL_ATTAINMENT_LEVELS}
\alias{EDUCATIONAL_ATTAINMENT_LEVELS}
\title{Census labels for different levels of educational achievement}
\format{
A tibble with two columns
\describe{
\item{Detailed}{a factor with 8 levels}
\item{Broad}{a factor with 5 levels}
}
}
\source{
https://api.census.gov/data
}
\usage{
EDUCATIONAL_ATTAINMENT_LEVELS
}
\description{
The census breaks down educational attainment to different levels of detail
in different tables. For example, there are 8 levels in
\href{https://api.census.gov/data/2021/acs/acs5/groups/B15001.html}{\code{B15001}}, but
only 5 in the
\href{https://api.census.gov/data/2021/acs/acs5/groups/C15002H.html}{\verb{C15002*}}
tables. Each row in this table connects a detailed level with a broader one.
}
\keyword{datasets}
