% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EDUCATIONAL_ATTAINMENT_METADATA}
\alias{EDUCATIONAL_ATTAINMENT_METADATA}
\title{Census variables for different levels of educational achievement}
\format{
A list of three data frames

\strong{sex_age}
: A data frame with eight variables
\itemize{
\item group     : chr
\item index     : int
\item variable  : chr
\item Sex       : chr
\item Lower Age : int
\item Upper Age : int
\item Age       : chr
\item Education : Factor
}

\strong{race_ethnicity}
: A data frame with ten variables
\itemize{
\item group                 : chr
\item index                 : int
\item Sex                   : chr
\item Lower Age             : int
\item Upper Age             : int
\item Age                   : chr
\item Education             : Factor
\item Suffix                : chr
\item Census Race/Ethnicity : chr
\item variable              : chr
}

\strong{poverty}
: A data frame with six variables
\itemize{
\item group     : chr
\item index     : int
\item variable  : chr
\item Sex       : chr
\item Education : Factor
\item Poverty   : chr
\item Lower Age : int
\item Upper Age : int
\item Age       : chr
}
}
\source{
https://api.census.gov/data
}
\usage{
EDUCATIONAL_ATTAINMENT_METADATA
}
\description{
The census breaks down educational attainment to different levels of detail
in different tables. For example, there are 8 levels in
\href{https://api.census.gov/data/2021/acs/acs5/groups/B15001.html}{\code{B15001}}, but
only 5 in the
\href{https://api.census.gov/data/2021/acs/acs5/groups/C15002H.html}{\verb{C15002*}}
tables or
\href{https://api.census.gov/data/2021/acs/acs5/groups/B17003.html}{\code{B17003}}.
Each row in this table connects a detailed level with a broader one.
}
\keyword{datasets}
