% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EMPLOYMENT_STATUS_METADATA}
\alias{EMPLOYMENT_STATUS_METADATA}
\title{Factor values associated with specific rows within employment-related ACS tables}
\format{
\subsection{EMPLOYMENT_STATUS_METADATA}{

A data frame with 439 rows and 12 columns:
\describe{
\item{group}{<chr>}
\item{index}{<int>}
\item{variable}{<chr>}
\item{Sex}{<chr>}
\item{Lower Age}{<int>}
\item{Upper Age}{<int>}
\item{Labor Force}{<chr>}
\item{Enlistment}{<chr>}
\item{Employment}{<chr>}
\item{Suffix}{<chr>}
\item{Census Race/Ethnicity}{<chr>}
\item{Poverty}{<chr>}
}
}
}
\source{
api.census.gov/data/acs/acs5/groups.html
}
\usage{
EMPLOYMENT_STATUS_METADATA
}
\description{
Factor values associated with specific rows within employment-related ACS tables
}
\keyword{datasets}
