% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_range.R
\name{extract_range}
\alias{extract_range}
\title{Extract two numbers from a range in census variables}
\usage{
extract_range(.range_text, .label, .defaults = c(0L, 999L))
}
\arguments{
\item{.range_text}{\verb{<chr>} a column of text with at least one number}

\item{.label}{\verb{<chr>} part of the name of each output column}

\item{.defaults}{\verb{<int[2]?>} the replacements for non-matches. Defaults to 0 and 999}
}
\value{
a tibble with two integer columns, \verb{Upper \{.label\}} and \verb{Lower \{.label\}}
}
\description{
Extract two numbers from a range in census variables
}
