% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FAMILIES_WITH_CHILDREN_METADATA}
\alias{FAMILIES_WITH_CHILDREN_METADATA}
\title{Categorize ACS variables counting families by number of children}
\format{
\subsection{FAMILIES_WITH_CHILDREN_METADATA}{

A data frame with 12 rows and 6 columns
\describe{
\item{Group}{\if{html}{\out{<chr>}} The table, always "B11003"}
\item{Variable}{\if{html}{\out{<chr>}} The full variable name, e.g. "B11003_001E"}
\item{Index}{\if{html}{\out{<int>}} The row of this variable in the table}
\item{Adults}{\if{html}{\out{<chr>}} The adult(s) heading the household}
\item{Children under 6}{\if{html}{\out{<lgl>}} Are there any children under 6 in this family?}
\item{Childr 6-17}{\if{html}{\out{<lgl>}} Are there any children 6-17 in this family?}
}
}
}
\source{
https://api.census.gov/data/2022/acs/acs1/groups/B11003.html
}
\usage{
FAMILIES_WITH_CHILDREN_METADATA
}
\description{
These data come from table B11003, "FAMILY TYPE BY PRESENCE AND AGE OF OWN CHILDREN UNDER 18 YEARS"
}
\keyword{datasets}
