% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_acs5_race_ethnicity_trends.R
\name{fetch_acs5_race_ethnicity_trends}
\alias{fetch_acs5_race_ethnicity_trends}
\title{Grab data from table B03002, which has race/ethnicity counts by tract.}
\usage{
fetch_acs5_race_ethnicity_trends(state, county, years)
}
\arguments{
\item{state}{the 2-digit FIPS code for a state, e.g. \code{55L} for Wisconsin}

\item{county}{the 3-digit FIPS code for a county, e.g. \code{101L} for Racine}

\item{years}{an integer vector of years to query. 2009 is the earliest possible year}
}
\value{
a tibble with tract, vintage, year, race/ethnicity, and population data
}
\description{
Grab data from table B03002, which has race/ethnicity counts by tract.
}
