% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_acs5_tract_data.R
\name{fetch_acs5_tract_data}
\alias{fetch_acs5_tract_data}
\title{Pull data from the 5-year American Community Survey}
\usage{
fetch_acs5_tract_data(variables, state, county, years = 2009:2021L)
}
\arguments{
\item{variables}{a character vector of variables like "B03002_001E"}

\item{state}{the two-digit FIPS code for the state, like 55 for Wisconsin}

\item{county}{the three-digit FIPS code for the county, like 101 for Racine}

\item{years, }{optional, the years to fetch data for. Defaults to 2009-2021}
}
\value{
a tibble of year, vintage, tract, group, index, variable, and value
}
\description{
The default years span the entire offering of 5-year ACS data.
Your variables will probably not all be available for every year.
}
