% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Download a set of data from the US Census API}
\usage{
fetch_data(
  variables,
  year,
  for_geo,
  for_items,
  survey_type,
  table_or_survey_code,
  ...,
  use_key = TRUE
)
}
\arguments{
\item{variables}{<chr[]> a vector of variable names, like \code{"B01001_001E"}}

\item{year}{an integer year, e.g. \code{2021L}}

\item{for_geo}{\verb{<chr>} A Census geography like "us," "state," "tract," or "school district (unified)."}

\item{for_items}{<chr[]> one or more instances of \code{for_geo} desired, e.g. \code{"*"} or \code{"000200"}, passed on to \code{\link[=build_for_geographies]{build_for_geographies()}}}

\item{survey_type}{e.g. "acs" or "dec"}

\item{table_or_survey_code}{e.g. "acs5" or "pl"}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic dots}}> list of key-value pairs to pass to \code{\link[=build_in_geographies]{build_in_geographies()}}}

\item{use_key}{<lgl?> optional, should the query include a Census API key from the system environment. Defaults to \code{TRUE}}
}
\value{
a tibble with "Group", "Index", "Value", and "Year" fields, as well as one field for each geography.
}
\description{
Download a set of data from the US Census API
}
\seealso{
\code{\link[=build_api_url]{build_api_url()}}
}
