% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_decennial_block_pops_by_race.R
\name{fetch_decennial_block_pops_by_race}
\alias{fetch_decennial_block_pops_by_race}
\title{Get decennial census estimates of per-block populations by race/ethnicity}
\usage{
fetch_decennial_block_pops_by_race(state, county)
}
\arguments{
\item{state}{two-digit FIPS code for the state, e.g. 55L for Wisconsin}

\item{county}{three-digit FIPS code for the county within the state, e.g. 101L for Racine}
}
\value{
a \code{\link[tibble:tibble]{tibble}} with four columns:
\describe{
\item{Vintage}{The decennial year, currently 2000, 2010, or 2020}
\item{Race/Ethnicity}{OMB text descriptions of Race/Ethnicities}
\item{GEOID}{The full FIPS code for each census block}
\item{Population}{The number of people in that block in that year of that race/ethnicity}
}
}
\description{
This function retrieves data from tables that splits people into race and Hispanic/Latin\@ ethnicity.
}
