% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_metadata_table.R
\name{fetch_metadata_table}
\alias{fetch_metadata_table}
\title{Fetch glossary about a specific ACS data set from the Census API.}
\usage{
fetch_metadata_table(.info_type, .year, .year_span)
}
\arguments{
\item{.info_type}{One of "geography", "groups", or "variables".}

\item{.year}{An integer year between 2004 and the current year, inclusive.}

\item{.year_span}{Either 1, 3, or 5, depending upon the desired time resolution.}
}
\value{
A tibble.
}
\description{
This function downloads a large JSON object and parses it into a tibble.
}
\concept{glossary}
