% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GEOGRAPHY_HIERARCHY_METADATA}
\alias{GEOGRAPHY_HIERARCHY_METADATA}
\title{Geographic levels of organization for data from the US Census Bureau's API}
\format{
A tibble with four columns
\describe{
\item{Label}{A short, title-case description of a geographic level}
\item{Code}{The three-digit code for a level}
\item{Geography}{The lowercase term used to query the API for a level}
\item{Parent Geos}{A character vector of other Geographies that must be specified to query for this level.}
}
}
\usage{
GEOGRAPHY_HIERARCHY_METADATA
}
\description{
Geographic levels of organization for data from the US Census Bureau's API
}
\keyword{datasets}
