% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_ACS_GEOGRAPHIES}
\alias{GLOSSARY_OF_ACS_GEOGRAPHIES}
\title{Details about levels of geographic detail in the most recent American Community Series}
\format{
A tibble with five columns
\describe{
\item{Geographic Level}{The verbatim text that you must pass to the API when referring to this geographic level.}
\item{Containing Geographies}{One or more sets of containing geographies that \strong{must} be specified when querying for this level of geography}
\item{Wildcard Option}{Which containing geography, if any, can be set to a wildcard when pulling data for this geographical level.}
\item{ACS1}{A reference date if this level is available in the 1-year dataset, otherwise \code{NA}.}
\item{ACS5}{A reference date if this level is available in the 5-year dataset, otherwise \code{NA}.}
}
}
\usage{
GLOSSARY_OF_ACS_GEOGRAPHIES
}
\description{
In the context of ACS data, "geography" is a specific type of geographical
area, such as a region, state, county, reservation, tract, or block. Some
geographical levels, such as region or state, can be given by themselves.
Others, such as county or place, must be queried within a specific containing
geographic level (usually state). Very small geographic areas, like blocks,
may require several levels of containing geographies, such as state, county,
and tract. Some containing geographies may be given as wildcards. For
example, you could ask for county-level data from every state all at once by
specifying "*" instead of a FIPS code for the state.
}
\concept{GLOSSARY}
\keyword{datasets}
