% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_ACS_VARIABLES}
\alias{GLOSSARY_OF_ACS_VARIABLES}
\title{Details about variables in the most recent American Community Series}
\format{
A list with three items
\describe{
\item{Dataset}{Whether this variable pertains to the 1-year or 5-year dataset.}
\item{Group}{The code for the group, such as "B01001B" or "B25063"}
\item{Index}{The row number in the group for this variable, like 2 or 5.}
\item{Variable}{The full code for this variable, like "B01001B_002E" or "B25063_005E"}
\item{Details}{A vector of one or more strings describing what the variable actually represents.}
}
}
\usage{
GLOSSARY_OF_ACS_VARIABLES
}
\description{
In the context of ACS data, a "variable" is a single row of data from
running a query about a specific topic. For example, group B010001 reports population
numbers by sex and age, and group B25063 reports the number of housing units
in different rent brackets. Technically, there are 4 variables for each row:
the estimate, its margin of error, and annotations for each of those. In
practice, we usually just want the estimate..
}
\concept{GLOSSARY}
\keyword{datasets}
