% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_AGE_AND_SEX}
\alias{GLOSSARY_OF_AGE_AND_SEX}
\title{The fundamental demographics of age and sex from table \code{B01001}}
\format{
\subsection{GLOSSARY_OF_AGE_AND_SEX}{

A data frame with 49 rows and 6 columns.
\describe{
\item{group}{<chr> Always "B01001"}
\item{index}{<int> The row in the source table that this GLOSSARY row describes.}
\item{variable}{<chr> The full variable name, e.g. "B01001_001E"}
\item{Sex}{<chr> One of "All," "Male," or "Female"}
\item{Lower Age}{<int> inclusive, in [-999, 85]}
\item{Upper Age}{<int> inclusive, in [4, 999]}
}
}
}
\source{
https://api.census.gov/data/2022/acs/acs1/groups/B01001.html
}
\usage{
GLOSSARY_OF_AGE_AND_SEX
}
\description{
The age splits are more detailed in the whole-population table than they are
in the tables that pertain to specific racial identities. A future project
is to add a second part to the table that maps from those tables' rows to
age and sex.
}
\keyword{datasets}
