% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_AGGREGATE_HOUSEHOLD_INCOME}
\alias{GLOSSARY_OF_AGGREGATE_HOUSEHOLD_INCOME}
\title{Aggregate income values for households, broken down by race, age, and source}
\format{
\subsection{GLOSSARY_OF_AGGREGATE_HOUSEHOLD_INCOME}{

An object of class \code{spec_tbl_df/tbl_df/tbl/data.frame} with
24 rows and 6 columns
\describe{
\item{Group}{\verb{<chr>} the identification code of the item's table}
\item{Index}{\verb{<int>} the item's row in its table}
\item{Race/Ethnicity}{\verb{<chr>} the racial/ethnic identity of people described by the item}
\item{Source}{\verb{<chr>} the type of income, e.g. earnings, interest, or SNAP}
\item{Public}{\verb{<lgl>} TRUE if the income is from a government assistance program.}
\item{Lower Age}{\verb{<int>} the lowest age of householders described by the item}
\item{Upper Age}{\verb{<int>}the highest age of householders described by the item}
\item{Subtotal}{\verb{<chr>} whether the row involves all, race/ethnicity, age, or source subtotals},
\item{Atomic}{\verb{<lgl>} TRUE if the item is an atomic observation, not a subtotal}
}
}
}
\source{
https://api.census.gov/data/2024/acs/acs5/groups.html
}
\usage{
GLOSSARY_OF_AGGREGATE_HOUSEHOLD_INCOME
}
\description{
These data come from 20 different tables. All of the tables, except for
B19050, which reports income by age of householder, have a single row, the
total income across all households in the geography. Tables B19025.* contain
values by race, and tables B19060-70 contain values by the nature of the
income.
}
\keyword{datasets}
