% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_CHILD_AGE_AND_FAMILY_STRUCTURE}
\alias{GLOSSARY_OF_CHILD_AGE_AND_FAMILY_STRUCTURE}
\title{Census variables for how many children of different ages live in families}
\format{
An object of class \code{tbl_df/tbl/data.frame} with
20 rows and 7 columns
\describe{
\item{Group}{\verb{<chr>} Always "B09002"}
\item{Index}{\verb{<int>}The row number in the table}
\item{Variable}{\verb{<chr>} The full census variable}
\item{Married}{\verb{<lgl>} Whether the family is single-parent or married}
\item{\verb{Sex of Householder}}{\verb{<chr>} Female, Male, or NA}
\item{\verb{Lower Age}}{\verb{<int>} inclusive from 0 to 17}
\item{\verb{Upper Age}}{\verb{<int>} exclusive from 0 to 17}
}
}
\usage{
GLOSSARY_OF_CHILD_AGE_AND_FAMILY_STRUCTURE
}
\description{
This table only lists "own children," so presumably that excludes foster
care and multi-generational homes with older householders.
}
\keyword{datasets}
