% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_CHILDREN_PER_FAMILY}
\alias{GLOSSARY_OF_CHILDREN_PER_FAMILY}
\title{Categorize ACS variables about children per family}
\format{
\subsection{GLOSSARY_OF_CHILDREN_PER_FAMILY}{

A data frame with 15 rows and 6 columns
\describe{
\item{Group}{<chr> The table, always "B09002"}
\item{Variable}{<chr> The full variable name, e.g. "B09002_001E"}
\item{Index}{<int> The row of this variable in the table}
\item{Adults}{<chr> The adult(s) heading the household}
\item{Lower Age}{<int> The age of the youngest children counted by this variable}
\item{Upper Age}{<int> The age of the oldest children counted by this variable}
}
}
}
\source{
https://api.census.gov/data/2022/acs/acs1/groups/B09002.html
}
\usage{
GLOSSARY_OF_CHILDREN_PER_FAMILY
}
\description{
These data come from table B17026, "OWN CHILDREN UNDER 18 YEARS BY FAMILY
TYPE AND AGE"
}
\keyword{datasets}
