% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_EDUCATION_AND_EMPLOYMENT}
\alias{GLOSSARY_OF_EDUCATION_AND_EMPLOYMENT}
\title{Crossed counts of educational attainment and employment status}
\format{
\subsection{GLOSSARY_OF_EDUCATION_AND_EMPLOYMENT}{

A data frame with 29 rows and 7 columns
\describe{
\item{Group}{\verb{<chr>} Always "B23006"}
\item{Index}{\verb{<int>}The row number in the table}
\item{Variable}{\verb{<chr>} The variable name for the population estimate}
\item{Education}{\verb{<chr>}one of five broad levels of educational attainment}
\item{\verb{Labor Force}}{\verb{<lgl>} \code{NA} means "All"}
\item{Civilian}{\verb{<lgl>} \code{NA} means "All"}
\item{Employed}{\verb{<lgl>} \code{NA} means "All"}
}
}
}
\source{
https://api.census.gov/data/2024/acs/acs5/groups/B23006.html
}
\usage{
GLOSSARY_OF_EDUCATION_AND_EMPLOYMENT
}
\description{
This table, \code{B23006}, uses the five-level attainment classification in the
"Broad" column of \code{\link{EDUCATIONAL_ATTAINMENT_LEVELS}}. Its
employment status information does not distinguish between full- and
part-time employment.
}
\keyword{datasets}
