% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_EMPLOYMENT_STATUS}
\alias{GLOSSARY_OF_EMPLOYMENT_STATUS}
\title{Factor values associated with specific rows within employment-related ACS tables}
\format{
\subsection{GLOSSARY_OF_EMPLOYMENT_STATUS}{

A data frame with 439 rows and 12 columns:
\describe{
\item{group}{\verb{<chr>}}
\item{index}{\verb{<int>}}
\item{variable}{\verb{<chr>}}
\item{Sex}{\verb{<chr>}}
\item{Lower Age}{\verb{<int>}}
\item{Upper Age}{\verb{<int>}}
\item{Labor Force}{\verb{<chr>}}
\item{Enlistment}{\verb{<chr>}}
\item{Employment}{\verb{<chr>}}
\item{Suffix}{\verb{<chr>}}
\item{Census Race/Ethnicity}{\verb{<chr>}}
\item{Poverty}{\verb{<chr>}}
}
}
}
\source{
https://api.census.gov/data/acs/acs5/groups.html
}
\usage{
GLOSSARY_OF_EMPLOYMENT_STATUS
}
\description{
Factor values associated with specific rows within employment-related ACS tables
}
\keyword{datasets}
