% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_FAMILIES_WITH_CHILDREN}
\alias{GLOSSARY_OF_FAMILIES_WITH_CHILDREN}
\title{Categorize ACS variables counting families by number of children}
\format{
\subsection{GLOSSARY_OF_FAMILIES_WITH_CHILDREN}{

A data frame with 12 rows and 6 columns
\describe{
\item{Group}{<chr> The table, always "B11003"}
\item{Variable}{<chr> The full variable name, e.g. "B11003_001E"}
\item{Index}{<int> The row of this variable in the table}
\item{Adults}{<chr> The adult(s) heading the household}
\item{Children under 6}{<lgl> Are there any children under 6 in this family?}
\item{Childr 6-17}{<lgl> Are there any children 6-17 in this family?}
}
}
}
\source{
https://api.census.gov/data/2022/acs/acs1/groups/B11003.html
}
\usage{
GLOSSARY_OF_FAMILIES_WITH_CHILDREN
}
\description{
These data come from table B11003, "FAMILY TYPE BY PRESENCE AND AGE OF OWN CHILDREN UNDER 18 YEARS"
}
\keyword{datasets}
