% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_FAMILY_STRUCTURE_AND_POVERTY}
\alias{GLOSSARY_OF_FAMILY_STRUCTURE_AND_POVERTY}
\title{Poverty status related to number and type of parents/guardians and children}
\format{
\subsection{GLOSSARY_OF_FAMILY_STRUCTURE_AND_POVERTY}{

A data frame with 41 rows and 7 columns
\describe{
\item{Index}{\if{html}{\out{<int>}} the row number}
\item{\verb{Below Poverty Level}}{\if{html}{\out{<lgl>}} TRUE if the families' incomes are below the federal poverty level}
\item{Married}{\if{html}{\out{<lgl>}} The Census keeps track about this I guess?}
\item{Male}{\if{html}{\out{<lgl>}} A male parent is present}
\item{Female}{\if{html}{\out{<lgl>}} A female parent is present}
\item{\verb{Under 5}}{\if{html}{\out{<lgl>}} At least one child under 5 is present}
\item{\verb{Over 4 Under 18}}{\if{html}{\out{<lgl>}} At least on child aged 5-17 is present}
}
}
}
\source{
https://api.census.gov/data/2024/acs/acs5/groups/B17010.html
}
\usage{
GLOSSARY_OF_FAMILY_STRUCTURE_AND_POVERTY
}
\description{
These data come from tables \verb{B17010*}, which have the same structure for all
racial identities and data sources.
}
\keyword{datasets}
