% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_POVERTY_FAMILY_AND_CHILDREN}
\alias{GLOSSARY_OF_POVERTY_FAMILY_AND_CHILDREN}
\title{Number of children by poverty status related type of parents/guardians.}
\format{
\subsection{GLOSSARY_OF_POVERTY_FAMILY_AND_CHILDREN}{

A data frame with 29 rows and 9 columns
\describe{
\item{Group}{\verb{<chr>} Always "B17006"}
\item{Index}{\verb{<int>} The row number.}
\item{Variable}{\verb{<chr>} computed from Group and Index.}
\item{\verb{Federal Poverty}}{\verb{<lgl>} TRUE if the family's income is below the Federal poverty level.}
\item{Married}{\verb{<lgl>} TRUE if the Census thinks the family's householders are wedded to one another}
\item{\verb{Male Parent}}{\verb{<lgl>} TRUE if there is a male parent/guardian present in the household}
\item{\verb{Female Parent}}{\verb{<lgl>} TRUE if there is a female parent/guarian present in the household}
\item{\verb{Lower Age}}{\verb{<int>} The youngest age in the range counted by this row}
\item{\verb{Upper Age}}{\verb{<int>} The oldest age in the range counted by this row}
}
}
}
\source{
https://api.census.gov/data/2024/acs/acs5/groups/B17006.html
}
\usage{
GLOSSARY_OF_POVERTY_FAMILY_AND_CHILDREN
}
\description{
These data come from table \code{B17006}.
}
\keyword{datasets}
