% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_STANDARD_OF_LIVING}
\alias{GLOSSARY_OF_STANDARD_OF_LIVING}
\title{Categorize ACS variables about income : poverty level ratios by family sustainability}
\format{
\subsection{GLOSSARY_OF_STANDARD_OF_LIVING}{

A data frame with 13 rows and 6 columns
\describe{
\item{Group}{\verb{<chr>} The table, always "B17026"}
\item{Variable}{\verb{<chr>} The full variable name, e.g. "B17026_001E"}
\item{Index}{\verb{<int>} The row of this variable in the table}
\item{Least Poverty Ratio}{\verb{<dbl>} The lowest ratio of income to poverty level in this tier}
\item{Greatest Poverty Ratio}{\verb{<dbl>} The highest ratio of income to poverty level in this tier}
\item{Standard of Living}{\verb{<chr>} One of "Everyone," "Unsustainable," or "Self-sustaining"}
}
}
}
\source{
https://api.census.gov/data/2022/acs/acs1/groups/B17026.html
}
\usage{
GLOSSARY_OF_STANDARD_OF_LIVING
}
\description{
These data come from table B17026, "RATIO OF INCOME TO POVERTY LEVEL OF
FAMILIES IN THE PAST 12 MONTHS."
A family-sustaining wage is widely considered to be three times the federal
poverty level.
}
\keyword{datasets}
