% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_TENURE_BY_AGE}
\alias{GLOSSARY_OF_TENURE_BY_AGE}
\title{Households by age of householder and owner or renter status}
\format{
\subsection{GLOSSARY_OF_TENURE_BY_AGE}{

A data frame with 21 rows and 5 columns
\describe{
\item{Group}{\if{html}{\out{<chr>}} Always "B25007"}
\item{Index}{\if{html}{\out{<int>}} Each row in the table}
\item{Variable}{\if{html}{\out{<chr>}} The full name for each estimate variable}
\item{Tenure}{\if{html}{\out{<chr>}} One of "All," "Owner," or "Renter"} "", "Owner occupied", "Owner occupied", "Owner occupied", "Owner occupied", "Owner occupied", "…
\item{Lower Age}{\if{html}{\out{<int>}} One of 15, 25, 35, 45, 55, 60, 75, or 85}
\item{Upper Age}{\if{html}{\out{<int>}} One of 24, 34, 44, 54, 59, 64, 74, 84, or 999}
}
}
}
\usage{
GLOSSARY_OF_TENURE_BY_AGE
}
\description{
Households by age of householder and owner or renter status
}
\keyword{datasets}
