% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_WORK_STATUS}
\alias{GLOSSARY_OF_WORK_STATUS}
\title{ACS data on age and full- or part-time employment}
\format{
\subsection{GLOSSARY_OF_WORK_STATUS}{

an object of class \code{tbl_df/tbl/data.frame} with
36 rows and
7 columns
\describe{
\item{Index}{\verb{<int>} the row in the census table}
\item{Variable}{\verb{<chr>} the full variable name}
\item{Lower Age}{\verb{<int>} the youngest age that this row counts}
\item{Upper Age}{\verb{<int>} the oldest age that this row counts}
\item{Employed}{\verb{<lgl>} \code{TRUE} if employed, \code{NA} if either status}
\item{Full Time}{\verb{<lgl>``TRUE} if full-time, \code{NA} if either status}
\item{Employment}{\verb{<chr>} a three-level factor with levels "Unemployed", "Part-Time", and "Full-Time"}
}
}
}
\source{
https://api.census.gov/2024/acs/acs5/groups/B23027.html
}
\usage{
GLOSSARY_OF_WORK_STATUS
}
\description{
ACS data on age and full- or part-time employment
}
\keyword{datasets}
