% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoist_table_glossary.R
\name{hoist_table_glossary}
\alias{hoist_table_glossary}
\title{Hoist details about a Census data group from a list of strings to separate columns}
\usage{
hoist_table_glossary(.glossary, .group, .fields)
}
\arguments{
\item{.glossary}{<data-frame> a table of glossary about Census API variables, e.g. \link{GLOSSARY_OF_ACS_VARIABLES}}

\item{.group}{<chr> the code for the group, e.g. "B01001"}

\item{.fields}{<chr[]> names for the new columns, e.g. c("Sex", "Age")}
}
\value{
a new data frame with more columns
}
\description{
Hoist details about a Census data group from a list of strings to separate columns
}
\examples{
hoist_table_glossary(GLOSSARY_OF_ACS_VARIABLES, "B06002", c("Median age", "Place of birth"))
}
\concept{glossary}
