% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_url.R
\name{info_url}
\alias{info_url}
\title{Get the URL for JSON metadata about one table of ACS data}
\usage{
info_url(.info_type, .year, .year_span)
}
\arguments{
\item{.info_type}{One of "geography", "groups", or "variables".}

\item{.year}{An integer year between 2004 and the current year, inclusive.}

\item{.year_span}{Either 1, 3, or 5, depending upon the desired time resolution.}
}
\value{
A string that contains a URL.
}
\description{
This will be a complete URL, including protocol and file extension, for
downloading metadata about geographies, groups of variables, or specific
variables.
}
\examples{
info_url("groups", 2021L, 5L)
}
