% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_counts_and_groups.R
\name{join_counts_and_groups}
\alias{join_counts_and_groups}
\title{Unite the grouping variables that define them to the counts from the ACS}
\usage{
join_counts_and_groups(acs_counts, acs_groups)
}
\arguments{
\item{acs_counts}{A data frame of counts per "Table" and "Row"}

\item{acs_groups}{A data frame of grouping variables for "Table" and "Row"}
}
\value{
A data frame where the counts and grouping variables are joined
}
\description{
Unite the grouping variables that define them to the counts from the ACS
}
\examples{
TBD <- NULL
}
