% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_list_to_frame.R
\name{json_list_to_frame}
\alias{json_list_to_frame}
\title{Convert a list of json objects from the census into a data frame}
\usage{
json_list_to_frame(.json_list)
}
\arguments{
\item{.json_list}{the list of objects}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}} with column names taken from the list
}
\description{
Convert a list of json objects from the census into a data frame
}
