% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_acs_metadata.R
\name{latest_acs_metadata}
\alias{latest_acs_metadata}
\title{Download glossary about the most recent versions of the 1- and 5-year ACS}
\usage{
latest_acs_metadata(information_type)
}
\arguments{
\item{information_type}{<chr> "geographies", "groups", or "variables"}
}
\value{
a tibble with dimensions that depend upon \code{information_type}
}
\description{
Download glossary about the most recent versions of the 1- and 5-year ACS
}
\concept{glossary}
