% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_other_geos.R
\name{make_other_geos}
\alias{make_other_geos}
\title{Possibly helpful function for creating a list of arguments for querying the Census API}
\usage{
make_other_geos(geography, ...)
}
\arguments{
\item{geography}{the name or numeric code of the target census geographic level}

\item{...}{name-value pairs for specific containing geographies, e.g. \code{state = 55L}}
}
\value{
a list of geographic levels and values
}
\description{
Possibly helpful function for creating a list of arguments for querying the Census API
}
\examples{
all_racine_tracts <- make_other_geos("tract", state = 55L, county = 101L)
}
