% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins_by_demography_and_place.R
\name{margins_by_demography_and_place}
\alias{margins_by_demography_and_place}
\title{Determine the relative contribution of different demographic groups to the populations of a set of areas.}
\usage{
margins_by_demography_and_place(
  block_data,
  ...,
  demography_field = "Race/Ethnicity",
  place_field = "Place",
  weight_field = "Population"
)
}
\arguments{
\item{block_data}{a tibble of census data that connects blocks to enclosing geographic areas.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> grouping fields in the data besides \code{demography} and \code{place}}

\item{demography_field}{optional, the field that identifies peoples' demographic group. Defaults to "Race/Ethnicity"}

\item{place_field}{optional, the field that denotes your custom geographic areas. Defaults to "Place"}

\item{weight_field}{optional, the field that denotes counts. Defaults to "Population"}
}
\value{
a tibble with proportions for estimating per-demographic values from per-tract data
}
\description{
Determine the relative contribution of different demographic groups to the populations of a set of areas.
}
