% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METADATA_ACS5}
\alias{METADATA_ACS5}
\title{Group, Variable, and Geography information for the latest 5-Year ACS}
\format{
A list with three items
\describe{
\item{geography}{Available geographies and how to ask for them in the API.}
\item{groups}{Each group's ID, its Universe, and its Description.}
\item{variables}{Each variable's Group, its full code, and Details about what it encodes.}
}
}
\usage{
METADATA_ACS5
}
\description{
Group, Variable, and Geography information for the latest 5-Year ACS
}
\concept{metadata}
\keyword{datasets}
