% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_recent_vintage.R
\name{most_recent_vintage}
\alias{most_recent_vintage}
\title{Query the Census API for the most recent release year of a dataset.}
\usage{
most_recent_vintage(survey_type, table_or_survey_code)
}
\arguments{
\item{survey_type}{e.g. "acs" or "dec"}

\item{table_or_survey_code}{e.g. "acs5" or "pl"}
}
\value{
an integer, probably at least 2024
}
\description{
Query the Census API for the most recent release year of a dataset.
}
