% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_acs_table_variable.R
\name{parse_acs_table_variable}
\alias{parse_acs_table_variable}
\title{Break an ACS "Variable" value into its three components.}
\usage{
parse_acs_table_variable(.x, col = "Variable")
}
\arguments{
\item{.x}{data.frame}

\item{col}{character}
}
\value{
data.frame
}
\description{
Break an ACS "Variable" value into its three components.
}
\examples{
parse_acs_table_variable(data.frame(Variable = "ABC123_456E"))
}
