% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RACE_ETHNICITY_SUBTABLE_METADATA}
\alias{RACE_ETHNICITY_SUBTABLE_METADATA}
\title{Race and ethnicity codes and labels used by the U.S. Census Bureau}
\format{
A tibble with ten rows and three columns:
\describe{
\item{Census Race}{A short description of a race or ethnicity}
\item{Suffix}{A capital letter designating a race or ethnicity}
\item{Inclusive Table}{The name of a table that includes people of single and multi-racial identities that claim this identity}
}
}
\usage{
RACE_ETHNICITY_SUBTABLE_METADATA
}
\description{
Race and ethnicity codes and labels used by the U.S. Census Bureau
}
\keyword{datasets}
