% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RACE_ETHNICITY_SUBTABLES}
\alias{RACE_ETHNICITY_SUBTABLES}
\title{Race and ethnicity codes and labels used by the U.S. Census Bureau}
\format{
A tibble with ten rows and five columns:
\describe{
\item{Census Race}{A short description of a race or ethnicity}
\item{Suffix}{A capital letter designating a race or ethnicity}
\item{Inclusive Table}{The name of a table that includes people of single and multi-racial identities that claim this identity}
\item{non-Hispanic}{The row in table \code{B03002} with counts of people with this racial identity and no Hispanic ethnic identity}
\item{Hispanic}{The row in table \code{B03002} with counts of people with this racial identity and a Hispanic ethnic identity}
}
}
\usage{
RACE_ETHNICITY_SUBTABLES
}
\description{
Race and ethnicity codes and labels used by the U.S. Census Bureau
}
\keyword{datasets}
