% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RACE_OR_ETHNIC_GROUP_TYPES}
\alias{RACE_OR_ETHNIC_GROUP_TYPES}
\title{For selected tables, an alphabetic suffix follows to indicate that a table is repeated for the nine major race and Hispanic or Latino groups:}
\format{
\subsection{RACE_OR_ETHNIC_GROUP_TYPES}{

A data frame with 9 rows and 2 columns.
\describe{
\item{Code}{<chr> A one-letter suffix}
\item{Race or Ethnic Group}{<chr> A racial or ethnic identity such as "Black Alone" or "Hispanic or Latino".}
}
}
}
\source{
https://www.census.gov/programs-surveys/acs/data/data-tables/table-ids-explained.html
}
\usage{
RACE_OR_ETHNIC_GROUP_TYPES
}
\description{
For selected tables, an alphabetic suffix follows to indicate that a table is repeated for the nine major race and Hispanic or Latino groups:
}
\keyword{datasets}
