% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{race_or_ethnicity}
\alias{race_or_ethnicity}
\title{Race and ethnicity codes and labels used by the U.S. Census Bureau}
\format{
A tibble with two columns:
\describe{
\item{code}{A capital letter designating a race or ethnicity}
\item{label}{A short description of a race or ethnicity}
}
}
\usage{
race_or_ethnicity
}
\description{
Race and ethnicity codes and labels used by the U.S. Census Bureau
}
\keyword{datasets}
