% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragged_list_to_tibble.R
\name{ragged_list_to_tibble}
\alias{ragged_list_to_tibble}
\title{Coerce a ragged list to a tibble with some list columns}
\usage{
ragged_list_to_tibble(.l)
}
\arguments{
\item{.l}{the ragged list}
}
\value{
a tibble
}
\description{
Coerce a ragged list to a tibble with some list columns
}
\examples{
rags <- list(One = list(A = 1:2, B = "a"), Two = list(A = 1))
ragged_list_to_tibble(rags)
}
