% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_geographies_with_geoid.R
\name{replace_geographies_with_geoid}
\alias{replace_geographies_with_geoid}
\title{Condense geographical hierarchy fields into one \code{"GEOID"} field.}
\usage{
replace_geographies_with_geoid(.x)
}
\arguments{
\item{.x}{a tibble from calling \code{\link[=fetch_data]{fetch_data}}}
}
\value{
a tibble with geographical fields replaced by their concatenation as \code{"GEOID"}
}
\description{
Condense geographical hierarchy fields into one \code{"GEOID"} field.
}
\examples{
tibble::tribble(
~ state, ~ county, ~ tract,  ~ Group,  ~ Index,  ~ Value, ~ Year,
55L,     101L,     "000200", "B01001", 1L,        5078L,   2021L
) |> replace_geographies_with_geoid()
}
