% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rusd_c23002a}
\alias{rusd_c23002a}
\title{Employment in the area served by Racine Unified School District, WI in 2019}
\format{
A list with 4 items:
\describe{
\item{url}{the actual URL used to query the Census's data API}
\item{json}{The JSON-formatted string that the Census responded with}
\item{list}{The list produced by \code{jsonlite::parse_json()}}
\item{data}{A data frame produced by \code{acs_counts_list_to_tibble()}}
}
}
\source{
https://www.census.gov/data/developers/guidance/api-user-guide.html
}
\usage{
rusd_c23002a
}
\description{
A list containing several stages of parsing results from querying the US
Census Bureau's data API for the American Community Survey. This set contains
the results from searching for summary table "C23002A" for the Racine Unified
School district (code 12360) in Wisconsin (55). It contains counts for white
people broken down by sex, age, and employment status.
}
\keyword{datasets}
