% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_in_glossary.R
\name{search_in_glossary}
\alias{search_in_glossary}
\title{Search for terms in a glossary in case-insensitive ways}
\usage{
search_in_glossary(.glossary, ...)
}
\arguments{
\item{.glossary}{the glo}

\item{...}{<\link[rlang:dyn-dots]{rlang::dyn-dots}> Named arrays of search terms.}
}
\value{
<dfr> a filtered version of \code{.glossary}
}
\description{
Each of the dynamic arguments to this function must be a named array of
strings.
The function will look for each string in the column of \code{.glossary} that
corresponds to the argument's name. For example, providing the argument
\code{Universe = c("Wom[ae]n", "years?")} will search the "Universe" column for any
elements that include both "women"  or "woman" and "year" or "years".
The search is NOT case-sensitive.
}
\examples{
hercacstables::search_in_glossary(
  hercacstables::GLOSSARY_OF_ACS_GEOGRAPHIES,
  `Wildcard Option` = "."
)
}
