% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SUBJECT_TYPES}
\alias{SUBJECT_TYPES}
\title{A code for the broad subject that a Census table covers.}
\format{
\subsection{TABLE_TYPES}{

A data frame with 31 rows and 2 columns.
\describe{
\item{Subject Number}{<chr> The zero-padded numeric code for this subject}
\item{Subject Name}{<chr> A very short description of the subject.}
}
}
}
\source{
https://www.census.gov/programs-surveys/acs/data/data-tables/table-ids-explained.html
}
\usage{
SUBJECT_TYPES
}
\description{
The second 2 characters identify the subject of the table.
These only apply to B, C, K20, S, R, and GCT Tables.
Profiles DP, NP, CP, and S0201 cover multiple topics, so they do not have any
characters to indicate a subject.
}
\keyword{datasets}
