% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_parts_from_whole.R
\name{subtract_parts_from_whole}
\alias{subtract_parts_from_whole}
\title{Computes a remainder values from measures of some subgroups and one all-group}
\usage{
subtract_parts_from_whole(
  .data_frame,
  grouping_column,
  value_column,
  whole_name,
  part_names,
  remainder_name
)
}
\arguments{
\item{.data_frame}{\if{html}{\out{<dataframe>}} the input data}

\item{grouping_column}{\if{html}{\out{<chr>}} the name of the column that defines the grouping}

\item{value_column}{\if{html}{\out{<chr>}} the name of the column that holds the values}

\item{whole_name}{\if{html}{\out{<chr>}} the level of the grouping that signifies the total}

\item{part_names}{<chr[]> the other levels of the grouping}

\item{remainder_name}{\if{html}{\out{<chr>}} the new name for the difference between whole and parts}
}
\value{
a new data frame with the same number of columns but extra rows and a new group level
}
\description{
Computes a remainder values from measures of some subgroups and one all-group
}
