% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TABLE_TYPES}
\alias{TABLE_TYPES}
\title{A Census table's ID always starts with an alphanumeric code for its type.}
\format{
\subsection{TABLE_TYPES}{

A data frame with 11 rows and 3 columns.
\describe{
\item{Table ID Characters}{<chr> The alphanumeric code for this type of table}
\item{Type of Table}{<chr> A human-readable name for the type of table.}
\item{Contains}{<chr> A description of what information this type of table will contain.}
}
}
}
\source{
https://www.census.gov/programs-surveys/acs/data/data-tables/table-ids-explained.html
}
\usage{
TABLE_TYPES
}
\description{
A Census table's ID always starts with an alphanumeric code for its type.
}
\keyword{datasets}
