% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_year_right_now.R
\name{the_year_right_now}
\alias{the_year_right_now}
\title{Today's year in the Gregorian calendar.}
\usage{
the_year_right_now()
}
\value{
\if{html}{\out{<int>}} probably 2024 or later
}
\description{
Today's year in the Gregorian calendar.
}
\examples{
the_year_right_now()

}
\seealso{
\code{\link[=as.POSIXlt]{as.POSIXlt()}}, \code{\link[=Sys.Date]{Sys.Date()}}
}
