% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TYPES_OF_RACE_OR_ETHNICITY}
\alias{TYPES_OF_RACE_OR_ETHNICITY}
\title{For selected tables, an alphabetic suffix follows to indicate that a table is repeated for the nine major race and Hispanic or Latino groups:}
\format{
\subsection{TYPES_OF_RACE_OR_ETHNICITY}{

A data frame with 9 rows and 2 columns.
\describe{
\item{Code}{\verb{<chr>} A one-letter suffix}
\item{Race or Ethnic Group}{\verb{<chr>} A racial or ethnic identity such as "Black Alone" or "Hispanic or Latino".}
}
}
}
\source{
https://www.census.gov/programs-surveys/acs/data/data-tables/table-ids-explained.html
}
\usage{
TYPES_OF_RACE_OR_ETHNICITY
}
\description{
For selected tables, an alphabetic suffix follows to indicate that a table is repeated for the nine major race and Hispanic or Latino groups:
}
\keyword{datasets}
