% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TYPES_OF_VARIABLE}
\alias{TYPES_OF_VARIABLE}
\title{ACS data contain variables ending in 1- or 2-letter codes.}
\format{
\subsection{TYPES_OF_VARIABLE}{

A data frame with 7 rows and 4 columns.
\describe{
\item{Variable Type}{\verb{<chr>} A 1- or 2-letter code}
\item{Title}{\verb{<chr>} A human-readable name for the type of variable.}
\item{Related Products}{\verb{<chr>} The types of Census data that contain this kind of variable.}
\item{Meaning}{\verb{<chr>} A detailed explanation of what this kind of variable represents.}
}
}
}
\source{
https://www.census.gov/data/developers/data-sets/acs-1year/notes-on-acs-api-variable-types.html
}
\usage{
TYPES_OF_VARIABLE
}
\description{
This table defines each variable type and its meaning.
}
\keyword{datasets}
