% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UNEMPLOYMENT_VARIABLES}
\alias{UNEMPLOYMENT_VARIABLES}
\title{Variables describing unemployment at two levels of age granularity}
\format{
\subsection{UNEMPLOYMENT_VARIABLES}{

A data frame with 140 rows and 7 columns
\describe{
\item{Sex}{\if{html}{\out{<chr>}} Female or Male}
\item{Lower Age}{\if{html}{\out{<int>}} 16 to 75}
\item{`Upper Age}{\if{html}{\out{<int>}} 19 to Inf}
\item{Labor Status}{\if{html}{\out{<chr>}} "In armed services", "Employed", "Unemployed", or "Not in labor force"}
\item{Group}{\if{html}{\out{<chr>}} "B23001" (more granular) or "C23001" (less granular)}
\item{Index}{\if{html}{\out{<int>}} The actual number of the variable in the ACS table.}
\item{Variable}{\if{html}{\out{<chr>}} The full ACS variable name for the estimate, e.g. "B23001_001E"}
}
}
}
\usage{
UNEMPLOYMENT_VARIABLES
}
\description{
Variables describing unemployment at two levels of age granularity
}
\keyword{datasets}
