% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_group_details.R
\name{unpack_group_details}
\alias{unpack_group_details}
\title{Pull apart the lists of details for one group of ACS data}
\usage{
unpack_group_details(.group)
}
\arguments{
\item{.group}{the code of the group, like "B25043" or "C15010B"}
}
\value{
a tibble with at least 4 columns, like \link{GLOSSARY_OF_ACS_VARIABLES}, but \code{Details} is split into multiple columns with uppercase letter names.
}
\description{
Pull apart the lists of details for one group of ACS data
}
\examples{
unpack_group_details("B01002B")

}
\seealso{
\link{GLOSSARY_OF_ACS_VARIABLES}
}
