% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{US_RACE_ETHNICITY_POPS_2022}
\alias{US_RACE_ETHNICITY_POPS_2022}
\title{US population size for 10 different race/ethnicity categories in 2022}
\format{
\subsection{US_RACE_ETHNICITY_POPS_2022}{

A data frame with 10 rows and 5 columns
\describe{
\item{us}{\if{html}{\out{<int>}} The geography used in the API call. Always 1}
\item{Group}{\if{html}{\out{<chr>}} The source table for the row, "B01001" with a suffix.}
\item{Index}{\if{html}{\out{<int>}} The row of this variable in the table. Always 1}
\item{Value}{\if{html}{\out{<int>}} The number of people in the race/ethnicity category.}
\item{Year}{\if{html}{\out{<int>}} The year used in the API call. Always 2022}
}
}
}
\source{
https://api.census.gov/data/2022/acs/acs5/groups.html
}
\usage{
US_RACE_ETHNICITY_POPS_2022
}
\description{
This is an example set that emulates the results of running
\code{\link[=fetch_data]{fetch_data()}} for two different sets of ACS tables.
The data for the "broad" level of detail are from tables B01001 its prefix-
designated sub-tables, such as B01001H for
\itemize{
\item labeled as "broad" in the Detail Level field.
\item which has population by age, with different tables for each identity
\item B03002 - which has population by race x ethnicity
}
}
\details{
The data from the B01001 tables are all from the first row, which is the
total population of that particular racial/ethnic identity.
The data from table B03002 are
}
\keyword{datasets}
