% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VARIABLE_TYPES}
\alias{VARIABLE_TYPES}
\title{ACS data contain variables ending in 1- or 2-letter codes.}
\format{
\subsection{VARIABLE_TYPES}{

A data frame with 7 rows and 4 columns.
\describe{
\item{Variable Type}{<chr> A 1- or 2-letter code}
\item{Title}{<chr> A human-readable name for the type of variable.}
\item{Related Products}{<chr> The types of Census data that contain this kind of variable.}
\item{Meaning}{<chr> A detailed explanation of what this kind of variable represents.}
}
}
}
\source{
https://www.census.gov/data/developers/data-sets/acs-1year/notes-on-acs-api-variable-types.html
}
\usage{
VARIABLE_TYPES
}
\description{
This table defines each variable type and its meaning.
}
\keyword{datasets}
