% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_in_group.R
\name{variables_in_group}
\alias{variables_in_group}
\title{Extract each variable for one group of 5-year ACS data}
\usage{
variables_in_group(.group)
}
\arguments{
\item{.group}{one ACS group, e.g. "B21005"}
}
\value{
a tibble with x columns
}
\description{
Extract each variable for one group of 5-year ACS data
}
