% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_api_url.R
\name{build_api_url}
\alias{build_api_url}
\title{Create an API call to send to api.census.gov}
\usage{
build_api_url(
  variables,
  for_geo,
  for_items,
  survey_type,
  table_or_survey_code,
  year,
  ...,
  use_key = TRUE
)
}
\arguments{
\item{variables}{a vector of variable names, like \code{"B01001_001E"}}

\item{for_geo}{the geographical level the data will describe, e.g. \code{"tract"}}

\item{for_items}{the specific instances of \code{for_geo} desired, e.g. \code{"*"} or \code{"000200"}}

\item{survey_type}{e.g. "acs" or "dec"}

\item{table_or_survey_code}{e.g. "acs5" or "pl"}

\item{year}{an integer year, e.g. \code{2021L}}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic dots}}> other items to pass to the query}

\item{use_key}{optional, should the query include a Census API key from the system environment. Defaults to \code{TRUE}}
}
\value{
one URL, as a string
}
\description{
Create an API call to send to api.census.gov
}
\examples{
hercacstables:::build_api_url(paste0("B25003_00", 1:3, "E"),
                              "tract",
                              "*",
                              "acs",
                              "acs5",
                              2020L,
                              state = 55L,
                              county = 101L,
                              use_key = FALSE)

hercacstables:::build_api_url(paste0("P1_00", c(1, 3, 4), "N"),
                              "tract",
                              "*",
                              "dec",
                              "pl",
                              2020L,
                              state = 55L,
                              county = 101L,
                              use_key = FALSE)
}
