% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_api_variable.R
\name{build_api_variable}
\alias{build_api_variable}
\title{Vectorized Creation of Census ACS Variables}
\usage{
build_api_variable(
  group_code,
  item_number,
  race_code = "",
  separator = "_",
  suffix = "E"
)
}
\arguments{
\item{group_code}{the group code, like "B18101"}

\item{item_number}{some integer between 0 and 999}

\item{race_code}{optional, either an empty string or one of [A..I]}

\item{separator}{optional, usually "_" for ACS variables.}

\item{suffix}{optional, usually "E" for ACS variables, but sometimes "N" for decennial data.}
}
\value{
a vector of strings, each of which is a variable name
}
\description{
All detail group variables have the format "B000OO*_000E", where the "B" and
"E" are literal and the "*" is a race code.
}
\examples{
groups <- c("B19013", "B18101", "B18101")
races <- c("", "", "I")
numbers <- 1
build_api_variable(groups, numbers, races)
}
