% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EMPLOYMENT_STATUS_METADATA}
\alias{EMPLOYMENT_STATUS_METADATA}
\title{Factor values associated with specific rows within employment-related ACS tables}
\format{
\subsection{EMPLOYMENT_STATUS_METADATA}{

A data frame with 439 rows and 12 columns:
\describe{
\item{group}{\if{html}{\out{<chr>}}}
\item{index}{\if{html}{\out{<int>}}}
\item{variable}{\if{html}{\out{<chr>}}}
\item{Sex}{\if{html}{\out{<chr>}}}
\item{Lower Age}{\if{html}{\out{<int>}}}
\item{Upper Age}{\if{html}{\out{<int>}}}
\item{Labor Force}{\if{html}{\out{<chr>}}}
\item{Enlistment}{\if{html}{\out{<chr>}}}
\item{Employment}{\if{html}{\out{<chr>}}}
\item{Suffix}{\if{html}{\out{<chr>}}}
\item{Census Race/Ethnicity}{\if{html}{\out{<chr>}}}
\item{Poverty}{\if{html}{\out{<chr>}}}
}
}
}
\source{
\link{api.census.gov/data/acs/acs5/groups.html}
}
\usage{
EMPLOYMENT_STATUS_METADATA
}
\description{
Factor values associated with specific rows within employment-related ACS tables
}
\keyword{datasets}
