% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Download a set of data from the US Census API}
\usage{
fetch_data(.variables, .year, .for_geo, .for_items, .other_geos, ...)
}
\arguments{
\item{.variables}{a vector of variable names, like \code{"B01001_001E"}}

\item{.year}{an integer year, e.g. \code{2021L}}

\item{.for_geo}{the geographical level the data will describe, e.g. \code{"tract"}}

\item{.for_items}{the specific instances of \code{.for_geo} desired, e.g. \code{"*"} or \code{"000200"}}

\item{.other_geos}{a named list of other geos, e.g. \code{list(state = 55L, county = 101L)}}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> further arguments for <\code{\link[=api_url]{api_url}}>.}
}
\value{
a tibble with "Variable", "Index", and "Value" fields, as well as one field for each geography.
}
\description{
Download a set of data from the US Census API
}
