% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_decennial_pops_by_race.R
\name{fetch_decennial_pops_by_race}
\alias{fetch_decennial_pops_by_race}
\title{Get decennial census estimates of per-block populations by race/ethnicity}
\usage{
fetch_decennial_pops_by_race(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=build_api_url]{build_api_url}}
  \describe{
    \item{\code{for_items}}{<chr[]> one or more instances of \code{for_geo} desired, e.g. \code{"*"} or \code{"000200"}, passed on to \code{\link[=build_for_geographies]{build_for_geographies()}}}
    \item{\code{use_key}}{<lgl?> optional, should the query include a Census API key from the system environment. Defaults to \code{TRUE}}
    \item{\code{for_geo}}{<chr> the geographical level the data will describe, e.g. \code{"tract"}}
  }}
}
\value{
a \code{\link[tibble:tibble]{tibble}} with at least four columns:
\describe{
\item{Vintage}{The decennial year, currently 2000, 2010, or 2020}
\item{Race/Ethnicity}{OMB text descriptions of Race/Ethnicities}
\item{...}{additional columns that identify a geographic unit}
\item{Population}{The number of people in that block in that year of that race/ethnicity}
}
}
\description{
This function retrieves data from tables that split people into race and
Hispanic/Latin\@ ethnicity.
}
