% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_json_as_list.R
\name{fetch_json_as_list}
\alias{fetch_json_as_list}
\title{Make a call to the Census API and convert the JSON response to an R list.}
\usage{
fetch_json_as_list(
  variables,
  year,
  for_geo,
  for_items,
  survey_type,
  table_or_survey_code,
  other_geos = NULL,
  ...,
  use_key = TRUE
)
}
\arguments{
\item{variables}{a vector of variable names, like \code{"B01001_001E"}}

\item{year}{an integer year, e.g. \code{2021L}}

\item{for_geo}{the geographical level the data will describe, e.g. \code{"tract"}}

\item{for_items}{the specific instances of \code{for_geo} desired, e.g. \code{"*"} or \code{"000200"}}

\item{survey_type}{e.g. "acs" or "dec"}

\item{table_or_survey_code}{e.g. "acs5" or "pl"}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic dots}}> other items to pass to the query}

\item{use_key}{optional, should the query include a Census API key from the system environment. Defaults to \code{TRUE}}
}
\value{
a list of items read from json
}
\description{
Make a call to the Census API and convert the JSON response to an R list.
}
\seealso{
\code{\link[=api_url]{api_url()}}

\code{\link[jsonlite:read_json]{jsonlite::read_json()}}
}
