% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_metadata_table.R
\name{fetch_metadata_table}
\alias{fetch_metadata_table}
\title{Fetch metadata about a specific ACS data set from the Census API.}
\usage{
fetch_metadata_table(.info_type, .year, .year_span)
}
\arguments{
\item{.info_type}{See \code{info_url}.}

\item{.year}{See \code{info_url}.}

\item{.year_span}{See \code{info_url}.}
}
\value{
A tibble.
}
\description{
This function downloads a large JSON object and parses it into a tibble.
}
\seealso{
\code{info_url}
}
