% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_AGE_AND_SEX}
\alias{GLOSSARY_OF_AGE_AND_SEX}
\title{The fundamental demographics of age and sex from tables \code{B01001[ A-H]}}
\format{
\subsection{GLOSSARY_OF_AGE_AND_SEX}{

A data frame with 80 rows and 5 columns.
\describe{
\item{Suffix}{\verb{<lgl>} TRUE for race-specific tables and FALSE for the all-races table.}
\item{Index}{\verb{<int>} The row in the source table that this GLOSSARY row describes.}
\item{Sex}{\verb{<chr>} One of "All," "Male," or "Female"}
\item{Lower Age}{\verb{<int>} the least age in the range, inclusive}
\item{Upper Age}{\verb{<int>} the greatest age in the range, inclusive}
}
}
}
\source{
https://api.census.gov/data/2022/acs/acs1/groups/B01001.html
}
\usage{
GLOSSARY_OF_AGE_AND_SEX
}
\description{
The age splits are more detailed in the whole-population table than they are
in the tables that pertain to specific racial identities.
}
\keyword{datasets}
