% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_EMPLOYMENT_STATUS}
\alias{GLOSSARY_OF_EMPLOYMENT_STATUS}
\title{Factor values associated with specific rows within employment-related ACS tables}
\format{
\subsection{GLOSSARY_OF_EMPLOYMENT_STATUS}{

A data frame with 439 rows and 12 columns:
\describe{
\item{group}{<chr>}
\item{index}{<int>}
\item{variable}{<chr>}
\item{Sex}{<chr>}
\item{Lower Age}{<int>}
\item{Upper Age}{<int>}
\item{Labor Force}{<chr>}
\item{Enlistment}{<chr>}
\item{Employment}{<chr>}
\item{Suffix}{<chr>}
\item{Census Race/Ethnicity}{<chr>}
\item{Poverty}{<chr>}
}
}
}
\source{
api.census.gov/data/acs/acs5/groups.html
}
\usage{
GLOSSARY_OF_EMPLOYMENT_STATUS
}
\description{
Factor values associated with specific rows within employment-related ACS tables
}
\keyword{datasets}
