% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_LABOR_FORCE_PARTICIPATION}
\alias{GLOSSARY_OF_LABOR_FORCE_PARTICIPATION}
\title{ACS data on sex, age, and labor force status}
\format{
\subsection{GLOSSARY_OF_LABOR_FORCE_PARTICIPATION}{

A data frame with 172 rows and 10 columns
\describe{
\item{Dataset}{"ACS1" or "ACS5"}
\item{Group}{Always "B23001"}
\item{Index}{<int> the row number of the variable}
\item{Variable{the full variable}
\item{Sex}{"Male" or "Female"}
\item{Lower Age}{<int> between 16 and 75, inclusive}
\item{Upper Age}{<int> between 19 and 999, inclusive}
\item{Labor Force}{<lgl>}
\item{Civilian}{<lgl>}
\item{Employed"}{<lgl>}
}
}
}
\usage{
GLOSSARY_OF_LABOR_FORCE_PARTICIPATION
}
\description{
ACS data on sex, age, and labor force status
}
\keyword{datasets}
