% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_LABOR_FORCE_PARTICIPATION}
\alias{GLOSSARY_OF_LABOR_FORCE_PARTICIPATION}
\title{ACS data on sex, age, and labor force status}
\format{
\subsection{GLOSSARY_OF_LABOR_FORCE_PARTICIPATION}{

A data frame with 172 rows and 10 columns
\describe{
\item{Dataset}{\verb{<chr>} "ACS1" or "ACS5"}
\item{Group}{\verb{<chr>} Always "B23001"}
\item{Index}{\verb{<int>} the row number of the variable}
\item{Variable}{\verb{<chr>} the full variable}
\item{Sex}{\verb{<chr>} "Male" or "Female"}
\item{Lower Age}{\verb{<int>} between 16 and 75, inclusive}
\item{Upper Age}{\verb{<int>} between 19 and 999, inclusive}
\item{Labor Force}{\verb{<lgl>}}
\item{Civilian}{\verb{<lgl>}}
\item{Employed"}{\verb{<lgl>}}
}
}
}
\usage{
GLOSSARY_OF_LABOR_FORCE_PARTICIPATION
}
\description{
ACS data on sex, age, and labor force status
}
\keyword{datasets}
