% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_SEX_BY_INCOME}
\alias{GLOSSARY_OF_SEX_BY_INCOME}
\title{Counts of people in income brackets, by sex and employment status.}
\format{
\subsection{GLOSSARY_OF_SEX_BY_INCEM}{

A data frame with 43 rows and 5 columns.
\describe{
\item{Index}{\verb{<int>} The row in the source table that this GLOSSARY row describes.}
\item{Sex}{\verb{<chr>} One of NA, "Male," or "Female"}
\item{Full-time}{\verb{<lgl>} \code{TRUE} if employed full-time for the past year.}
\item{Lower Bound}{\verb{<dbl>} the lowest income in the range, inclusive}
\item{Upper Bound}{\verb{<dbl>} the greatest income in the range, inclusive}
}
}
}
\source{
https://api.census.gov/data/2024/acs/acs1/groups/B19325.html

https://api.census.gov/data/2024/acs/acs1/groups/B20005.html
}
\usage{
GLOSSARY_OF_SEX_BY_INCOME
}
\description{
The rows in this glossary apply to both table \code{B19325} AND \code{B20005}. The
difference is that \code{B19325} deals with all income and \code{B20005} just earnings.
}
\keyword{datasets}
