% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_TENURE_DURATIONS}
\alias{GLOSSARY_OF_TENURE_DURATIONS}
\title{People in households by the length of their current residence}
\format{
\subsection{GLOSSARY_OF_TENURE_BY_AGE}{

A data frame with 36 rows and 8 columns
\describe{
\item{Group}{\verb{<chr>} Always "B25026"}
\item{Index}{\verb{<int>} The row number from the table}
\item{Variable}{\verb{<chr>} The full code needed to query the API for this value}
\item{Tenure}{\verb{<chr>} Renter or Owner occupied}
\item{Shortest Duration}{\verb{<int>} The fewest years of residency in this count}
\item{Longest Duration}{\verb{<int>} The most years of residency in this count}
}
}
}
\source{
https://api.census.gov/data/2024/acs/acs5/groups/B25026.html
}
\usage{
GLOSSARY_OF_TENURE_DURATIONS
}
\description{
A note of warning! The data for this table come from the biennial
\href{https://www.census.gov/programs-surveys/ahs}{American Housing Survey},
so you should only use data from even-numbered years.
}
\keyword{datasets}
