% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hercacstables-package.R
\docType{package}
\name{hercacstables-package}
\alias{hercacstables-package}
\title{hercacstables: Work with American Community Survey Tables through the US Census API}
\description{
The American Community Survey (ACS) from the US Census's website returns data in a weirdly idiosyncratic way. There are many subtotals in each table. Grouping variables are organized in a tree-like, rather than tabular fashion. This package is intended to make it easy to access and use the ACS data with R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://higherx4racine.github.io/hercacstables/}
}

}
\author{
\strong{Maintainer}: Ben Taft \email{ben.taft@career2cradle.org}

}
\keyword{internal}
