% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_api.R
\name{make_census_file_regex}
\alias{make_census_file_regex}
\title{Use look up fields to build a glob to search for downloaded ACS data.}
\usage{
make_census_file_regex(code, acs_type, is_metadata = FALSE)
}
\arguments{
\item{code}{this is a string containing table and (maybe?) year info}

\item{acs_type}{an integer, usually 1 or 5, but sometimes 3}

\item{is_metadata}{logical}
}
\value{
a file name as a string
}
\description{
Use look up fields to build a glob to search for downloaded ACS data.
}
\examples{
make_census_file_regex("C23002A", 5)
}
