% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_acs_table_variable.R
\name{parse_acs_table_variable}
\alias{parse_acs_table_variable}
\title{Break an ACS "Variable" value into its three components.}
\usage{
parse_acs_table_variable(.x, col = "Variable")
}
\arguments{
\item{.x}{tibble}

\item{col}{character}
}
\value{
tibble
}
\description{
Break an ACS "Variable" value into its three components.
}
\examples{
parse_acs_table_variable(tibble(Variable = "ABC123_456E"))
#> # A tibble: 1 x 3
#>   Table    Row `Statistic Type`
#>   <chr>  <int> <chr>
#> 1 ABC123   456 E
}
