% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_and_separate.R
\name{pivot_and_separate}
\alias{pivot_and_separate}
\title{Take a wide-formatted response from the Census API and convert it to long}
\usage{
pivot_and_separate(.frame, ...)
}
\arguments{
\item{.frame}{a data frame from the census with columns named like \code{B01001_001E}
@param ... <\code{\link[rlang:dyn-dots]{dynamic-dots}}> named regular expressions for extracting fields from the column names}
}
\value{
a pivoted data frames where the columns that matched the regular expression are converted to rows that are described by the extracted fields.
}
\description{
Take a wide-formatted response from the Census API and convert it to long
}
\seealso{
\code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}

\code{\link[tidyr:separate_wider_delim]{tidyr::separate_wider_regex()}}
}
